+++ internal/api/handler/sales.go
@@ -48,6 +48,7 @@ func (h *Handler) CalculateTotalSales(c *gin.Context) {
 // @Produce json
 // @Security ApiKeyAuth
 // @Param Sale body entity.Sale true "Sale data"
+// @Param branch_id header string true "Branch ID"
 // @Success 201 {object} products.SaleResponse
 // @Failure 400 {object} products.Error
 // @Failure 500 {object} products.Error
@@ -96,6 +97,13 @@ func (h *Handler) CreateSales(c *gin.Context) {
 	req.SoldBy = c.MustGet("id").(string)
 	req.CompanyId = companyID
 
+	branchId := c.GetHeader("branch_id")
+	if branchId == "" {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Branch ID is required in the header"})
+		return
+	}
+	req.BranchId = branchId
+
 	res, err := h.ProductClient.CreateSales(c, &req)
 	if err != nil {
 		h.log.Error("Error creating sale", "error", err.Error())
@@ -115,6 +123,7 @@ func (h *Handler) CreateSales(c *gin.Context) {
 // @Security ApiKeyAuth
 // @Param id path string true "Sale ID"
 // @Param Sale body entity.SaleUpdate true "Updated sale data"
+// @Param branch_id header string true "Branch ID"
 // @Success 200 {object} products.SaleResponse
 // @Failure 400 {object} products.Error
 // @Failure 500 {object} products.Error
@@ -131,6 +140,14 @@ func (h *Handler) UpdateSales(c *gin.Context) {
 	}
 
 	req.CompanyId = c.MustGet("company_id").(string)
+
+	branchId := c.GetHeader("branch_id")
+	if branchId == "" {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Branch ID is required in the header"})
+		return
+	}
+	req.BranchId = branchId
+
 	res, err := h.ProductClient.UpdateSales(c, &req)
 	if err != nil {
 		h.log.Error("Error updating sale", "error", err.Error())
@@ -149,13 +166,26 @@ func (h *Handler) UpdateSales(c *gin.Context) {
 // @Produce json
 // @Security ApiKeyAuth
 // @Param id path string true "Sale ID"
+// @Param branch_id header string true "Branch ID"
 // @Success 200 {object} products.SaleResponse
 // @Failure 400 {object} products.Error
 // @Failure 500 {object} products.Error
 // @Router /sales/{id} [get]
 func (h *Handler) GetSales(c *gin.Context) {
 	id := c.Param("id")
-	req := &products.SaleID{Id: id, CompanyId: c.MustGet("company_id").(string)}
+
+	// Получаем branch_id из заголовков
+	branchId := c.GetHeader("branch_id")
+	if branchId == "" {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Branch ID is required in the header"})
+		return
+	}
+
+	req := &products.SaleID{
+		Id:        id,
+		CompanyId: c.MustGet("company_id").(string),
+		BranchId:  branchId,
+	}
 
 	res, err := h.ProductClient.GetSales(c, req)
 	if err != nil {
@@ -175,6 +205,7 @@ func (h *Handler) GetSales(c *gin.Context) {
 // @Produce json
 // @Security ApiKeyAuth
 // @Param filter query entity.SaleFilter false "Filter parameters"
+// @Param branch_id header string true "Branch ID"
 // @Success 200 {object} products.SaleList
 // @Failure 400 {object} products.Error
 // @Failure 500 {object} products.Error
@@ -182,14 +213,24 @@ func (h *Handler) GetSales(c *gin.Context) {
 func (h *Handler) GetListSales(c *gin.Context) {
 	var filter products.SaleFilter
 
+	// Преобразуем параметры фильтра из запроса
 	if err := c.ShouldBindQuery(&filter); err != nil {
 		h.log.Error("Error parsing SaleFilter", "error", err.Error())
 		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
 		return
 	}
-	filter.CompanyId = c.MustGet("company_id").(string)
 
-	// Fetch the list of sales
+	// Получаем branch_id из заголовков
+	branchId := c.GetHeader("branch_id")
+	if branchId == "" {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Branch ID is required in the header"})
+		return
+	}
+
+	filter.CompanyId = c.MustGet("company_id").(string)
+	filter.BranchId = branchId
+
+	// Получаем список продаж с учетом фильтрации
 	res, err := h.ProductClient.GetListSales(c, &filter)
 	if err != nil {
 		h.log.Error("Error retrieving sales list", "error", err.Error())
@@ -197,15 +238,14 @@ func (h *Handler) GetListSales(c *gin.Context) {
 		return
 	}
 
+	// Дополнительная информация о клиентах, продавцах и товарах
 	for i, sale := range res.Sales {
-
 		clientRes, err := h.UserClient.GetClient(c, &user.UserIDRequest{Id: sale.ClientId, CompanyId: c.MustGet("company_id").(string)})
 		if err == nil {
 			res.Sales[i].ClientName = clientRes.FullName
 			res.Sales[i].ClientPhoneNumber = clientRes.Phone
 		} else {
 			h.log.Error("Error fetching customer details", "customer_id", sale.ClientId, "error", err.Error())
-			log.Println("1", err)
 		}
 
 		supplier, err := h.UserClient.GetUser(c, &user.UserIDRequest{Id: sale.SoldBy, CompanyId: c.MustGet("company_id").(string)})
@@ -213,9 +253,9 @@ func (h *Handler) GetListSales(c *gin.Context) {
 			res.Sales[i].SoldByName = supplier.FirstName
 		} else {
 			h.log.Error("Error fetching customer details", "customer_id", sale.SoldBy, "error", err.Error())
-			log.Println("2", err)
 		}
 
+		// Информация о проданных товарах
 		for j, item := range sale.SoldProducts {
 			productRes, err := h.ProductClient.GetProduct(c, &products.GetProductRequest{
 				Id:        item.ProductId,
@@ -226,7 +266,6 @@ func (h *Handler) GetListSales(c *gin.Context) {
 				res.Sales[i].SoldProducts[j].ProductImage = productRes.ImageUrl
 			} else {
 				h.log.Error("Error fetching customer details", "customer_id", filter.CompanyId, "error", err.Error())
-				log.Println("3", err)
 			}
 		}
 	}
@@ -242,13 +281,25 @@ func (h *Handler) GetListSales(c *gin.Context) {
 // @Produce json
 // @Security ApiKeyAuth
 // @Param id path string true "Sale ID"
+// @Param branch_id header string true "Branch ID"
 // @Success 200 {object} products.Message
 // @Failure 400 {object} products.Error
 // @Failure 500 {object} products.Error
 // @Router /sales/{id} [delete]
 func (h *Handler) DeleteSales(c *gin.Context) {
 	id := c.Param("id")
-	req := &products.SaleID{Id: id, CompanyId: c.MustGet("company_id").(string)}
+
+	branchId := c.GetHeader("branch_id")
+	if branchId == "" {
+		c.JSON(http.StatusBadRequest, gin.H{"error": "Branch ID is required in the header"})
+		return
+	}
+
+	req := &products.SaleID{
+		Id:        id,
+		CompanyId: c.MustGet("company_id").(string),
+		BranchId:  branchId,
+	}
 
 	res, err := h.ProductClient.DeleteSales(c, req)
 	if err != nil {
